requirements.txt
Flask
opencv-python-headless
Pillow
numpy
gunicorn
Flask-Limiter==3.5.0
python-dotenv==1.0.0
